refactor(nft): Remove button for signing to generate code
